# docs/seo-brain/openapi.v1.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SEO Brain API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/ai/task-kinds": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Task Kinds",
    "operationId": "task_kinds_api_v1_ai_task_kinds_get",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Task Kinds Api V1 Ai Task Kinds Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/models": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Models",
    "operationId": "models_api_v1_ai_models_get",
    "parameters": [
     {
      "name": "provider_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Provider Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Models Api V1 Ai Models Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/models/sync": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Sync Models",
    "description": "Seed catalog defaults per provider (idempotent) and optionally discover models from the provider (read-only).",
    "operationId": "sync_models_api_v1_ai_models_sync_post",
    "parameters": [
     {
      "name": "provider_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Provider Id"
      }
     },
     {
      "name": "discover",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Discover"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Sync Models Api V1 Ai Models Sync Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/models/{mid}": {
   "patch": {
    "tags": [
     "ai"
    ],
    "summary": "Update Model",
    "operationId": "update_model_api_v1_ai_models__mid__patch",
    "parameters": [
     {
      "name": "mid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Mid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ModelUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Model Api V1 Ai Models  Mid  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/ai/usage": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Usage",
    "operationId": "usage_api_v1_ai_usage_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "group_by",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "model",
       "title": "Group By"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Usage Api V1 Ai Usage Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/budget": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Budget",
    "operationId": "budget_api_v1_ai_budget_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Budget Api V1 Ai Budget Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "ai"
    ],
    "summary": "Set Budget",
    "description": "Human-set monthly budget per site (site_settings key `ai`). Thresholds stay 80/100/120 %.",
    "operationId": "set_budget_api_v1_ai_budget_put",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BudgetSet"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Set Budget Api V1 Ai Budget Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/routing/preview": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Routing Preview",
    "description": "Why would this task go to which model — no call is made.",
    "operationId": "routing_preview_api_v1_ai_routing_preview_get",
    "parameters": [
     {
      "name": "task_kind",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Task Kind"
      }
     },
     {
      "name": "site_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     },
     {
      "name": "priority",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "normal",
       "title": "Priority"
      }
     },
     {
      "name": "provider",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Provider"
      }
     },
     {
      "name": "model",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Model"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Routing Preview Api V1 Ai Routing Preview Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/estimate": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Estimate",
    "operationId": "estimate_api_v1_ai_estimate_post",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EstimateBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Estimate Api V1 Ai Estimate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/prompts": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "List Prompts",
    "operationId": "list_prompts_api_v1_ai_prompts_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     },
     {
      "name": "scope",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Scope"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Prompts Api V1 Ai Prompts Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Create Prompt",
    "operationId": "create_prompt_api_v1_ai_prompts_post",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PromptCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Prompt Api V1 Ai Prompts Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/prompts/{pid}": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Get Prompt",
    "operationId": "get_prompt_api_v1_ai_prompts__pid__get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Pid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Prompt Api V1 Ai Prompts  Pid  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/prompts/{pid}/versions": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Add Version",
    "operationId": "add_version_api_v1_ai_prompts__pid__versions_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Pid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VersionCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add Version Api V1 Ai Prompts  Pid  Versions Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/prompts/versions/{vid}": {
   "patch": {
    "tags": [
     "ai"
    ],
    "summary": "Patch Version",
    "operationId": "patch_version_api_v1_ai_prompts_versions__vid__patch",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VersionPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Patch Version Api V1 Ai Prompts Versions  Vid  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ai/prompts/versions/{vid}/preview": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Preview Version",
    "description": "Rendered prompt with the site's MemoryPack — nothing is sent to a provider.",
    "operationId": "preview_version_api_v1_ai_prompts_versions__vid__preview_post",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PromptTestBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Preview Version Api V1 Ai Prompts Versions  Vid  Preview Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BudgetSet": {
    "properties": {
     "budget_usd_month": {
      "type": "number",
      "maximum": 100000.0,
      "exclusiveMinimum": 0.0,
      "title": "Budget Usd Month"
     }
    },
    "type": "object",
    "required": [
     "budget_usd_month"
    ],
    "title": "BudgetSet"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ModelUpdate": {
    "properties": {
     "display": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display"
     },
     "tier": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "fast",
         "balanced",
         "quality",
         "reasoning"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Tier"
     },
     "tags": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tags"
     },
     "context_tokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Context Tokens"
     },
     "price_in_per_m": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price In Per M"
     },
     "price_out_per_m": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price Out Per M"
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     }
    },
    "type": "object",
    "title": "ModelUpdate"
   },
...
  }
 }
}